requirements.txt
flask==3.0.0
werkzeug==3.0.1
mysql-connector-python==8.2.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
imbalanced-learn==0.11.0
python-dotenv==1.0.0
gunicorn==21.2.0
